First commit of in-construction transect generation tool
This commit introduces two files, which together allow users to generate
transects through MPAS-O, as well as any other ocean data where the lat,
lon, and depth of each data point is given in the correct numpy array
format.  Currently, MPAS and SOSE data are able to be ingested, plotted
in regular meridional transects around Antarctica, and plotted
(including a plot of the difference).  A brief description of the
methodology:

1) Load data, massage into correct form
2) Send data and coordinate data to transect generator, which:
 a) assesses extent of transect
 b) trims transect over land (so, for example, one can start a transect
    over land, in order to capture coastal ocean cells automatically)
 c) reprojects transect and ocean data points to a local stereographic
    (i.e. cartesian) grid
 d) horizontally interpolates data, and depth of data, to transect
points
 e) vertically interpolates data to regular vertical grid
3) plots resulting regularly-gridded transect data

The decision to actually generate regularly-gridded transect data allows
for quantitative comparisons between observations and model output.

At this point, these tools are not in MPAS-Analysis format, and have
only been tested on Wolf (LANL IC).  The intention of moving development
here, is to allow other MPAS-Analysis developers access to these
scripts, with the intention of developing them into official
MPAS-Analysis scripts (as well as continuing development of the transect
methodology itself).